Tighten recovery contracts
